Rotate screenpad 90 degrees
This commit addresses the fact that every linux distro in existence gets
the default configuration wrong on the ASUS Zenbook UX 534 FTC. In
particular, you wind up with the screenpad initialized in portrait mode
(turned sideways (which renders it unreadable) and un-mousable),
as opposed to landscape. This makes it very difficult to identify what's
going on in sway, for instance, when you Super-Shift-0 a window onto
that workspace and need to manipulate it further, mousing on that
display is crazy-town in addition to the sideways rendered at 4 point
font sizing or less.